Change the Session ID from int to long, because int has become too small.

+++ src/ehupatras/webrecommendation/A010MainClassDistanceMatrixEuclidean.java
@@ -12,7 +12,7 @@ public class A010MainClassDistanceMatrixEuclidean {
 	private Matrix m_matrix;
 	
 	public void createDistanceMatrix(String databaseWD,
-			ArrayList<Integer> sampleSessionIDs,
+			ArrayList<Long> sampleSessionIDs,
 			ArrayList<String[]> sequencesUHC,
 			float[][] roleWeights){
 		m_matrix = new SimilarityMatrixEuclidean(sampleSessionIDs);
@@ -69,7 +69,7 @@ public class A010MainClassDistanceMatrixEuclidean {
 		A001MainClassCreateDatabase database = new A001MainClassCreateDatabase();
 		//database.createDatabase(databaseWD);
 		database.loadDatabase(databaseWD);
-		ArrayList<Integer> sampleSessionIDs = database.getSessionsIDs();
+		ArrayList<Long> sampleSessionIDs = database.getSessionsIDs();
 		ArrayList<String[]> sequencesUHC = database.getInstantiatedSequences();
